+++ lib/src/document_form.dart
@@ -76,14 +76,15 @@ class _DocumentFormState extends State<DocumentForm> {
   }
 
   void _saveDocument(BuildContext context) {
-     formKey.currentState.save();
+    formKey.currentState.save();
     if (widget.document == null) {
-      Document doc = Document.fromMap(_documentValues, persistenceProvider: widget.documentList.persistenceProvider);
+      Document doc = Document.fromMap(_documentValues,
+          persistenceProvider: widget.documentList.persistenceProvider);
       widget.documentList.add(doc);
     } else {
       widget.document.updateValues(_documentValues);
     }
-    Navigator.pop(context); 
+    Navigator.pop(context);
   }
 
   @override